Works, mostly.
Since it all uses the same singular array,it has a habit of "Skipping" or "jumping" and having the fire act in unusual ways. Countered this by giving the fire "HP". However, this can be improved - save changes to a new array instead, that way, no concurrent modification issues will crop up.